Revert "Added backend functionality and update project"
