--- backend/controllers/authController.js
@@ -1,87 +1,0 @@
-const User = require('../models/user');
-const { hashPassword, comparePassword } = require('../helpers/auth');
-const jwt = require('jsonwebtoken');
-
-const test = (req, res) => {
-    res.json('test is working');
-}
-
-const registerUser = async (req, res) => {
-    try {
-        const {name, email, password} = req.body;
-        //check if user name exists
-        if(!name) {
-            return res.json({
-                error: 'Name is required'
-            })
-        };
-        //check if password exists
-        if(!password || password.length < 6){
-            return res.json({
-                error: 'Password is required and should at least 6 characters long'});
-        };
-        //check if email exists
-        const exist = await User.findOne({email});
-        if(exist) {
-            return res.json({
-                error: 'Email already exists'
-            });
-        };
-
-        //hash password
-        const hashedPassword = await hashPassword(password);
-        //Create user database
-        const user = await User.create({name, email, password: hashedPassword});
-        return res.json(user);
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-//Login Endpoint
-const loginUser = async (req, res) => {
-    try {
-        const {email, password} = req.body;
-        //check if user exists
-        const user = await User.findOne({email});
-        if(!user) {
-            return res.json({
-                error: 'Invalid User credentials'
-            });
-        };
-        //check if password is correct
-        const match = await comparePassword(password, user.password);
-        if(match) {
-            jwt.sign({email: user.email, id: user._id, name: user.name}, process.env.JWT_SECRET, {}, (err, token) => {
-                if(err) throw err;
-                res.cookie('token', token).json(user)
-            })
-        }
-        if(!match) {
-            res.json({
-                error: 'Wrong password'
-            });
-        };
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-const getProfile = async (req, res) => {
-    const{token} = req.cookies;
-    if(token) {
-        jwt.verify(token, process.env.JWT_SECRET, {}, (err, user) =>{
-            if(err) throw err;
-            res.json(user);
-        })
-    } else {
-        res.json(null);
-    }
-}
-
-module.exports = {
-    test,
-    registerUser,
-    loginUser,
-    getProfile
-}
