--- backend/helpers/auth.js
@@ -1,26 +1,0 @@
-const bcrypt = require('bcryptjs');
-
-const hashPassword = async (password) => {
-    return new Promise((resolve, reject) => {
-        bcrypt.genSalt(12, (err, salt) => {
-            if(err) {
-                reject(err);
-            }
-            bcrypt.hash(password, salt, (err, hash) => {
-                if(err) {
-                    reject(err);
-                }
-                resolve(hash);
-            });
-        });
-    });
-}
-
-const comparePassword = async (password, hashed) => {
-    return bcrypt.compare(password, hashed);
-}
-
-module.exports = {
-    hashPassword,
-    comparePassword
-}
