+++ src/App.jsx
@@ -1,28 +1,49 @@
-import {Route, createBrowserRouter, createRoutesFromElements, RouterProvider} from "react-router-dom";
-import HomePage from "./pages/HomePage";
-import MainLayout from "./layouts/MainLayout";
-import axios from "axios";
-import Login from "./pages/Login";
-import Register from "./pages/Register";
-import Dashboard from "./pages/Dashboard";
+import React from "react";
+import Navbar from "./components/Navbar/Navbar";
+import NavbarBanner from "./components/Navbar/NavbarBanner";
+import Hero from "./components/Hero/Hero";
+import NumberCounter from "./components/NumberCounter/NumberCounter";
+import WhyChooseUs from "./components/WhyChooseUs/WhyChooseUs";
+import Img1 from "./assets/banner1.png";
+import Img2 from "./assets/banner2.png";
+import Banner from "./components/Banner/Banner";
+import SubjectCard from "./components/SubjectCard/SubjectCard";
+import Testimonial from "./components/Testimonial/Testimonial";
+import Footer from "./components/Footer/Footer";
 
+const BannerData = {
+  image: Img1,
+  tag: "CUSTOMIZE WITH YOUR SCHEDULE",
+  title: "Personalized Professional Online Tutor on Your Schedule",
+  subtitle:
+    "Our scheduling system allows you to select based on your free time. Lorem ipsum demo text for template. Keep track of your students class and tutoring schedules, and never miss your lectures. The best online class scheduling system with easy accessibility.Lorem ipsum is a placeholder text commonly used to demonstrate the visual form",
+  link: "#",
+};
 
-axios.defaults.baseURL = 'http://localhost:8000';
-axios.defaults.withCredentials = true;
-
-const router = createBrowserRouter(
-  createRoutesFromElements(
-    <Route path = '/' element= {<MainLayout />}>
-      <Route index element={<HomePage />} />
-      <Route path = '/register' element={<Register />} />
-      <Route path = '/login' element={<Login />} />
-      <Route path = '/dashboard' element={<Dashboard />} />
-    </Route>
-  )
-);
+const BannerData2 = {
+  image: Img2,
+  tag: "CUSTOMIZE WITH YOUR SCHEDULE",
+  title: "Talented and Qualified Tutors to Serve You for Help",
+  subtitle:
+    "Our scheduling system allows you to select based on your free time. Lorem ipsum demo text for template. Keep track of your students class and tutoring schedules, and never miss your lectures. The best online class scheduling system with easy accessibility. Lorem ipsum is a placeholder text commonly used",
+  link: "#",
+};
 
 const App = () => {
-  return <RouterProvider router={router} />;
+  return (
+    <main className="overflow-x-hidden">
+      <Navbar />
+      <NavbarBanner />
+      <Hero />
+      <NumberCounter />
+      <WhyChooseUs />
+      <Banner {...BannerData} />
+      <Banner {...BannerData2} reverse={true} />
+      <SubjectCard />
+      <Testimonial />
+      <Footer />
+    </main>
+  );
 };
 
 export default App;
