+++ src/components/Banner/Banner.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { motion } from "framer-motion";
 import { SlideUp } from "../../utility/animation";
