+++ src/components/Footer/Footer.jsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { FaFacebook, FaInstagram, FaLinkedin } from "react-icons/fa";
 import { HiLocationMarker } from "react-icons/hi";
 //import { MdComputer } from "react-icons/md";
