+++ src/components/Navbar/Navbar.jsx
@@ -4,7 +4,6 @@ import { MdMenu } from "react-icons/md";
 import { motion } from "framer-motion";
 import CSCShivam from"../../assets/CSCshivam.png";
 import ResponsiveMenu from "./ResponsiveMenu.jsx";
-import { Link } from "react-router-dom";
 
 const Navbar = () => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -47,16 +46,10 @@ const Navbar = () => {
           </div>
           {/* CTA Button section */}
           <div className="hidden lg:block space-x-6">
-           <Link to="/login"
-           className="font-semibold"
-           >
-            Login
-            </Link> 
-            <Link to="/register"
-            className="text-white bg-secondary font-semibold rounded-full px-5 py-1 "
-            >
+            <button className="font-semibold">Sign in</button>
+            <button className="text-white bg-secondary font-semibold rounded-full px-5 py-1 ">
               Register
-              </Link>
+            </button>
           </div>
           {/* Mobile Hamburger Menu */}
           <div className="lg:hidden" onClick={() => setIsOpen(!isOpen)}>
