+++ src/components/Navbar/NavbarBanner.jsx
@@ -1,6 +1,5 @@
-
-import { motion } from "framer-motion";
 import React from "react";
+import { motion } from "framer-motion";
 
 const NavbarBanner = () => {
   const [isOpen, setIsOpen] = React.useState(true);
