+++ src/components/Testimonial/Testimonial.jsx
@@ -1,4 +1,4 @@
-
+import React from "react";
 import Slider from "react-slick";
 
 const TestimonialsData = [
