--- src/pages/Dashboard.jsx
@@ -1,13 +1,0 @@
-import { useContext } from "react";
-import { UserContext } from "../context/userContext";
-
-export default function Dashboard() {
-    const {user} = useContext(UserContext);
-    return (
-        <div>
-            <h1>Dashboard</h1>
-            {!!user && (<h2>Hello {user.name},</h2>)}
-            <h2>Welcome {user.name}</h2>
-        </div>
-    )
-}
